+++ src/api/cupom/criar_cupom.rs
@@ -4,6 +4,7 @@ use axum::{
 use serde::Deserialize;
 use uuid::Uuid;
 use std::sync::Arc;
+use rust_decimal::Decimal;
 use crate::{api::{AppState, dto::AppError}, models::Usuario};
 
 
@@ -12,8 +13,8 @@ pub struct CriarCupomRequest {
     pub codigo: String,
     pub descricao: String,
     pub tipo_desconto: String,
-    pub valor_desconto: Option<f64>,
-    pub valor_minimo: Option<f64>,
+    pub valor_desconto: Option<Decimal>,
+    pub valor_minimo: Option<Decimal>,
     pub data_validade: String,
     pub limite_uso: Option<i32>,
 }
