+++ src/api/marketing/atualizar_promocao.rs
@@ -1,6 +1,7 @@
 use axum::{
     Extension, Json, extract::{Path, State}, response::IntoResponse, http::StatusCode
 };
+use rust_decimal::Decimal;
 use serde::Deserialize;
 use uuid::Uuid;
 use std::sync::Arc;
@@ -16,8 +17,8 @@ pub struct AtualizarPromocaoRequest {
     pub nome: String,
     pub descricao: String,
     pub tipo_desconto: String,
-    pub valor_desconto: Option<f64>,
-    pub valor_minimo: Option<f64>,
+    pub valor_desconto: Option<Decimal>,
+    pub valor_minimo: Option<Decimal>,
     pub data_inicio: String,
     pub data_fim: String,
     pub dias_semana_validos: Option<Vec<u8>>,
